+++ x509-certificate/CHANGELOG.md
@@ -21,6 +21,8 @@ Released on ReleaseDate.
   `ring::pkcs8::Document`. (The PKCS#8 representation can now be obtained
   from the returned `InMemorySigningKeyPair`.)
 * `X509CertificateBuilder::new()` no longer accepts a `KeyAlgorithm` argument.
+* `X509CertificateBuilder` now implements `Default`. `new()` is deprecated
+  in favor of `default()`.
 * `X509CertificateBuilder::create_with_random_keypair()` now requires a
   `KeyAlgorithm` argument.
 * `X509CertificateBuilder` has gained a `create_with_key_pair()` method that